// controls/richtexteditor/blazor-script/editor-scripts/common/interface.ts
export interface IEditorModel {
    currentDocument?: Document
    execCommand?: Function
    observer?: Observer
    markdownSelection?: MarkdownSelection
    undoRedoManager?: UndoRedoManager | UndoRedoCommands
    nodeSelection?: NodeSelection
    mdSelectionFormats?: MDSelectionFormats
    domNode?: DOMNode
    nodeCutter?: NodeCutter
    formatPainterEditor?: IFormatPainterEditor
    domTree?: DOMMethods
    linkObj?: LinkCommand | MDLink
    videoObj?: VideoCommand
    audioObj?: AudioCommand
    imgObj?: ImageCommand
    formatPainterObj?: FormatPainterActions
    tableObj?: TableCommand | MDTable
    pasteObj?: PasteCleanupAction
    emojiPickerObj?: EmojiPickerAction
    editableElement?: Element
    userAgentData?: CustomUserAgentData
    destroy?(): void
    beforeSlashMenuApplyFormat?(): void
    codeBlockObj?: CodeBlockPlugin
}
/**
 * Interface for checklist item model
 */
export interface IChecklistItemModel {
    /** Defines the type of the checklist item */
    text: string;
    iconCss: string;
}
/**
 * @hidden
 * @deprecated
 */
export interface IDropDownItemModel extends DropDownItemModel {
    cssClass?: string
    command?: string
    subCommand?: string
    value?: string
    text?: string
}

/**
 * @hidden
 * @deprecated
 */
export interface IToolbarItemModel extends ItemModel {
    command?: string
    subCommand?: string
}

/**
 * Represents the details of an image integrated into the Rich Text Editor.
 */
export interface IImageCommandsArgs {
    /** Specifies the `src` attribute of the image. */
    url?: string
    /** Represents the current selection instance. */
    selection?: NodeSelection
    /** Specifies the minimum, maximum, and actual width of the image. */
    width?: { minWidth?: string | number; maxWidth?: string | number; width?: string | number }
    /** Specifies the minimum, maximum, and actual height of the image. */
    height?: { minHeight?: string | number; maxHeight?: string | number; height?: string | number }
    /** Describes the alternate text attribute for the image. */
    altText?: string
    /** Defines the CSS class names to be applied to the image. */
    cssClass?: string
    /** Refers to the image element that is to be edited. */
    selectParent?: Node[]
}
// ...
export interface ITouchData {
    prevClientX?: number
    prevClientY?: number
    clientX?: number
    clientY?: number
}
// ...
export interface ICodeBlockLanguageModel {
    label?: string
    language?: string
}

/**
 * Provides detailed information about an actionBegin event.
 */
export interface ActionBeginEventArgs {
    /** Specifies the type of the current action. */
    requestType?: string
    /** Indicates whether to cancel the current action. */
    cancel?: boolean
    /**
     * Specifies the current toolbar or dropdown item involved in the action.
     *
     * @deprecated
     */
    item?: IToolbarItemModel | IDropDownItemModel
    /** Specifies the event that initiated the action, such as mouse, keyboard, or drag events. */
    originalEvent?: MouseEvent | KeyboardEvent | DragEvent
    /** Specifies the name of the event. */
    name?: string
    /** Specifies whether the selection type is a dropdown. */
    selectType?: string
    /**
     * Provides details about URL actions.
     *
     * @deprecated
     */
    itemCollection?: IItemCollectionArgs
    /**
     * Defines the emoji picker details.
     *
     * @deprecated
     */
    emojiPickerArgs?: IEmojiPickerArgs
    /**
     * Defines the content to be exported.
     *
     * @deprecated
     */
    exportValue?: string
}

export interface IEmojiPickerArgs {
    emojiSettings: EmojiSettingsModel

}

/**
 * Provides detailed information about a Print event in the Rich Text Editor (RTE).
 */
export interface PrintEventArgs extends ActionBeginEventArgs {
    /** Defines the Rich Text Editor (RTE) element associated with the Print event. */
    element?: Element
}

/**
 * @deprecated
 */
export interface IItemCollectionArgs {
    /** Defines the instance of the current selection */
    selection?: NodeSelection
    /** Defines the HTML elements of currently selected content */
    selectNode?: Node[]
    /** Defines the parent HTML elements of current selection */
    selectParent?: Node[]
    /** Defines the URL action details for link element */
    url?: string
    /** Defines the Display Text action details for link element */
    text?: string
    /** Defines the title of the link action details */
    title?: string
    /** Defines the target as string for link element */
    target?: string
    /** Defines the element to be inserted */
    insertElement?: Element
}

/**
 * @deprecated
 */
export interface IExecutionGroup {
    command: string
    subCommand?: string
    value?: string
}

/**
 * Provides information about a notification event in the rich text editor.
 */
export interface NotifyArgs {
    module?: string
    args?: KeyboardEvent | MouseEvent | ClickEventArgs | ClipboardEvent | TouchEvent
    cancel?: boolean
    requestType?: string
    enable?: boolean
    properties?: object
    selection?: NodeSelection
    link?: HTMLInputElement
    selectNode?: Node[]
    selectParent?: Node[]
    url?: string
    text?: string
    isWordPaste?: boolean
    title?: string
    target?: string
    member?: string
    /** Specifies the name of the notifier handling the event. */
    name?: string
    /** Represents the range of text selection involved in the notification. */
    range?: Range
    /** Describes the action associated with the notification event. */
    action?: string
    callBack?(args?: string | IImageCommandsArgs, cropImageData?:
    { [key: string]: string | boolean | number }[], pasteTableSource?: string): void
    file?: Blob
    insertElement?: Element
    touchData?: ITouchData
    allowedStylePropertiesArray?: string[]
    isPlainPaste?: boolean
    formatPainterSettings?: FormatPainterSettingsModel
    ariaLabel?: string
    /**
     * Defines the source of the Table content.
     *
     * @private
     */
    pasteTableSource?: string
}
// ...
export interface ToolbarClickEventArgs {
    /**
     * Determines if the current toolbar click action can be canceled.
     */
    cancel: boolean
    /**
     * Defines the current Toolbar Item Object being clicked.
     */
    item: ItemModel
    /**
     * Contains the original mouse event arguments related to the toolbar click.
     */
    originalEvent: MouseEvent
    /**
     * Specifies the request type associated with the toolbar click event.
     */
    requestType: string
    /**
     * Specifies the name of the event.
     */
    name?: string
}
// ...
export interface IToolbarItems {
    template?: string
    tooltipText?: string
    command?: string
    subCommand?: string
    undo?: boolean
    click?: EmitType<ClickEventArgs>
}